Move common styling for dl into own component
