+++ client/app/queue/CaseSnapshot.jsx
@@ -1,10 +1,11 @@
-import { after, css, merge } from 'glamor';
+import { css } from 'glamor';
 import moment from 'moment';
 import PropTypes from 'prop-types';
 import React from 'react';
 import { connect } from 'react-redux';
 import _ from 'lodash';
 
+import CaseDetailsDescriptionList from './components/CaseDetailsDescriptionList';
 import SelectCheckoutFlowDropdown from './components/SelectCheckoutFlowDropdown';
 import JudgeStartCheckoutFlowDropdown from './components/JudgeStartCheckoutFlowDropdown';
 import COPY from '../../COPY.json';
@@ -25,20 +26,6 @@ const snapshotParentContainerStyling = css({
   '& > div:first-child': { paddingLeft: '3rem' },
 
   '& .Select': { maxWidth: '100%' }
-});
-
-const definitionListStyling = css({
-  margin: '0',
-  '& dt': merge(
-    after({ content: ':' }),
-    {
-      color: COLORS.GREY_MEDIUM,
-      float: 'left',
-      fontSize: '1.5rem',
-      marginRight: '0.5rem',
-      textTransform: 'uppercase'
-    }
-  )
 });
 
 const headingStyling = css({
@@ -125,19 +112,19 @@ export class CaseSnapshot extends React.PureComponent {
     return <div className="usa-grid" {...snapshotParentContainerStyling} {...snapshotChildResponsiveWrapFixStyling}>
       <div className="usa-width-one-fourth">
         <h3 {...headingStyling}>{COPY.CASE_SNAPSHOT_ABOUT_BOX_TITLE}</h3>
-        <dl {...definitionListStyling}>
+        <CaseDetailsDescriptionList>
           <dt>{COPY.CASE_SNAPSHOT_ABOUT_BOX_TYPE_LABEL}</dt>
           <dd>{renderAppealType(this.props.appeal)}</dd>
           <dt>{COPY.CASE_SNAPSHOT_ABOUT_BOX_DOCKET_NUMBER_LABEL}</dt>
           <dd>{appeal.docket_number}</dd>
           {this.daysSinceTaskAssignmentListItem()}
-        </dl>
+        </CaseDetailsDescriptionList>
       </div>
       <div className="usa-width-one-fourth">
         <h3 {...headingStyling}>{COPY.CASE_SNAPSHOT_TASK_ASSIGNMENT_BOX_TITLE}</h3>
-        <dl {...definitionListStyling}>
+        <CaseDetailsDescriptionList>
           {this.taskAssignmentListItems()}
-        </dl>
+        </CaseDetailsDescriptionList>
       </div>
       { this.props.featureToggles.phase_two &&
         this.props.loadedQueueAppealIds.includes(appeal.vacols_id) &&
